Update the CMake from upstream libxml2 repo
Use patch files rather than copying the build system in anticipation of a release.

Update the vcpkg code with the changes to the CMake. It no longer automatically detects libraries and then turns on support. Also enable threads as it doesn't seem like its possible to build without that support anymore and the vcpkg build options that were being used to turn off support weren't actually doing that.

The downloaded sources don't contain the xmlwin32version.h.in file so we have to still copy that.