Update libxml2 to v2.9.13
The GNOME libxml2 GitLab repository is now listing an official distribution tarball in the release section. Modify the download URL to use that one instead.

Remove the xmlwin32version.h.in file as it was removed from the repository.

Also add in a post 2.9.13 patch to fix the `LIBXML2_WITH_VALID=OFF` build.