Add boringssl port
Fleshes out a port of BoringSSL for WebKit use in order to enable HTTP/3. Targets the latest version referenced by ngtcp2 repositories which will be used as the backend for HTTP/3 in cURL.

Build modified to install targets and to toggle on and off testing and tooling.

Some additional patches were added that address compilation errors in MSVC 2022. Those can be removed at a later date.