Fixup CMake exports
Use `vcpkg_cmake_config_fixup` in the ports that generate CMake config files. Modify the install location for the CMake config files to be `${CMAKE_INSTALL_DATADIR}/${port}` which is the default for the function.

Add `cmake-vcpkg-webkit` which wraps the functions `vcpkg_cmake_config_fixup` and `vcpkg_fixup_pkgconfig`. If `VCPKG_DISABLE_CMAKE_FIXUP` is `ON` then `vcpkg_cmake_config_fixup` will delete the `*.cmake` files. If `VCPKG_DISABLE_PKGCONFIG_FIXUP` os `ON` then `vcpkg_fixup_pkgconfig` will delete the `pkgconfig` directories in `lib` and `debug/lib`. If they are not set then the functions behave normally. This is so a triplet can disable pkgconfig and CMake configs if necessary.

Remove any additional directories that were unused, and clean up the portfiles.