Prefer setting CONFIG_PATH for cmake fixup
A number of patches in the repository would modify the install paths to use `${CMAKE_INSTALL_DATADIR}` so it matched where `vcpkg` expected things to be. To simplify things drop these patches and call `vcpkg_cmake_config_fixup` with the proper `CONFIG_PATH`.

Modify the `vcpkg_cmake_config_fixup` wrapper to handle empty directories to prevent warnings during the build.